Add a method and API to suspend attacks.
Added an Admin package to hold and control data about the application itself. Added a Controller to configure the state via API.

The Attack class will now check if it's in a valid state before beginning an attack. The Health manager will also check if the system has been in a non-running state for excessive time.