Refactor ContainerManager into a much simpler method that only uses a set. The get method requires the class type as an argument, and returns that class type as a return value.
Main benefit is we prevent creation of extra Container objects and only need to compare a programmed Unique Identifier.

Decent amount of rework in Platform classes to use this and in unit tests to inherit things properly. Converted some unit tests to use Spring @Autowired and @MockBean via ContextConfiguration to safely load the ConfigurationManager.