Initial setup for using a Pipeline to push.
Added necessary gitignore files for doing work in VSCode.
Renamed the "test" stage to "develop", and the final stage to "master" in Dockerfile to match branch names.
Updated docker-compose accordingly.

Configured the Pipeline YML file to build from a variable branch, and target the Dockerfile target of the same name.