SCT-7354 Add a Shell Script Experiment to perform a Memory Burn
A shell script consumes and holds onto most of the memory of a container for an extended period.

It achieves this by finding the limit of available memory in the cgroup from /sys/fs/cgroup/memory/, and comparing it against the usage.

The difference, minus 1024*1024 bytes, is held by using `dd` to create a block of that size, and holding it as the STDIN of a long sleep command.

This results in a container running with 1MB of free memory.

This can result in some different errors.
- Orchestrator may be overprovisioned and cannot allocate new instances because memory limit is too high and being consumed here.
- Application may experience Out of Memory Errors and completely crash
- Application may experience Out of Memory Errors on threads serving complex requests, but not on the Liveness Probe