Added a test utilities module
This module does not use the project pom as a parent, and is imported in the project pom in the `test` scope. This allows any of its `main` resources to be `test` resources, including application.properties and logback-test.xml.

The module is installed early as part of the Dockerfile for dependency issues, and then explicitly deleted before the final image copies JARs via Wildcards.