Bug fix in the ExperimentManager that could cause an infinite loop if a forced experiment chose an EC2 platform with all containers stopped.
This involved moving some things from ExperimentManager to PlatformManager for simplicity, as well as cleaning up some unit tests affected as a result of changes.

Some of the unit tests did not use SpringJUnit4ClassRunner, so cleaned those up. Also cleaned up a lot of common mocks and used an @Before to hold common configuration.