Updated the terminateProcess script with 'Physical' and 'Container' appropriate actions
Using the symlink of /proc/1/exe and checking if it is systemd or init.

The physical branch kills several processes based on matching names. There has been room left to add more processes, and left in such a way that is git-friendly

The container variant just attempts to kill pid 1 with a SIGINT