SCT-10569 - Fix an issue where files in the classpath override scripts in the JAR.
Script Manager now filters out Resources that are Files.

It also searches for scripts in all classpath resources, not just the first match.

This required removing some unit tests from the ScriptManagerTest class, but they were mostly useless.

These tests are failing because, when run through Surefire, the files are not yet part of a JAR. The Script Manager filters them out.

The removed unit tests provided little to no value. The class depends on the final compiled structure. Maybe integration tests in the future can handle this.