Ignore extra paths related to MkDocs
The `public` and `.public` directories are used in the CI build of MkDocs.

The directories should be kept absent to ensure both that local builds of MkDocs are not sent to Git, and that the directories also aren't used for other content that will conflict with the Docs Build process.